chore: refactored container module
